+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 ### 功能构建
 
 - 优化部分代码，修复潜在的空指针异常。
+  - com.dwarfeng.familyhelper.plugin.finance.handler.pusher.FamilyhelperPusher。
+  - com.dwarfeng.familyhelper.plugin.life.handler.pusher.FamilyhelperPusher。
+  - com.dwarfeng.familyhelper.plugin.notify.handler.sender.EmailSenderRegistry。
   - com.dwarfeng.familyhelper.plugin.notify.handler.router.PermissionRouterRegistry。
 
 - settingrepo 模块功能新增。